Avoid spurious group add/remove notifications
When using `groups` in combination with `is_admin and/or
`oidc_login_default_group`, on each login, the user removed and
immediately re-added to the relevant groups (`admin` or the value of
`oidc_login_default_group`).

This is annoying, since it generates two notfications each time:

"An administrator removed you from group admin"
"An administrator added you to group admin"

To fix this, when `groups` is set, we first determine the full list of
groups to join and only then adjust the memberships.