webdav: Login user with token instead of password only
Previously, we'd login the WebDAV user only with a password.
However, this leads to issues, especially when using WebDAV (e.g.
with rclone) and the Nextcloud browser client in parallel. After
some time, the browser would refresh and log out the user. This
issue has been seen previously with the normal login flow as well
and was fixed in commit 3a50a43611d859daed44b065db86b16225052404.