Only cache JWKs and well-known HTTP OK responses
Previously, we would cache all responses from the IdP.
However, when the IdP or a reverse proxy in front of it
returns a non-200 code such as 500, we don't want to
cache the response. Caching the incorrect response would
prevent JWKs or the well-known configuration to be parsed
correctly and hence users from logging in.